feat: final changes, and add documentation

+++ back/src/entities/contacts.entitie.ts
@@ -18,7 +18,7 @@ class Contact {
   @Column({ type: "varchar", array: true, length: 120 })
   emails: string[];
 
-  @Column("int", { array: true })
+  @Column("bigint", { array: true })
   phone: number[];
 
   @CreateDateColumn({ type: "date" })
